requirements.txt
# NAYAM - Phase 1: Core Infrastructure & Data Spine
# ===================================================

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Validation & Serialization
pydantic==2.5.2
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.6

# File Handling
aiofiles==23.2.1

# Logging
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.23.2
httpx==0.25.2

# Environment
python-dotenv==1.0.0

# ===================================================
# NAYAM - Phase 2: Intelligence & Decision Support
# ===================================================

# AI / LLM
groq==0.25.0
openai==1.12.0
tiktoken==0.12.0

# Local STT (Whisper)
faster-whisper==1.2.1

# Document text extraction
PyPDF2==3.0.1
python-docx==1.1.2

# Vector math (cosine similarity in-process)
numpy==1.26.3

# Dense embeddings & FAISS vector search
faiss-cpu==1.13.2
sentence-transformers==5.2.3

# ===================================================
# NAYAM - Phase 3: Predictive Governance & Geo-Spatial
# ===================================================

# Geo-Spatial (PostGIS integration — optional at runtime)
geoalchemy2==0.15.0
shapely==2.0.6

# Encryption / Privacy Layer
cryptography==42.0.2

# Predictive / ML
scikit-learn==1.6.1

# ===================================================
# NAYAM - Phase 4: Offline-First & Production Readiness
# ===================================================

# Rate limiting
slowapi==0.1.9

# Structured logging
structlog==24.1.0

# Monitoring / health
prometheus-client==0.21.1

# Docker health probe support (ASGI)
async-timeout==5.0.1
